--- tests/test_build.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env sh
-
-g++ -Wall -Wextra -Wpedantic -std=c++17 -g $1 ../la.cpp ../strassen.cpp ../operators.cpp -o test.out
